requirements.txt
python-dotenv==1.1.0
slack_bolt==1.23.0
